# GODRIVER-2628 optimize allocations in selectServerFromDescription