GODRIVER-2628 

## Summary
Optimize allocations in selectServerFromDescription

## Background & Motivation
Same way as in #1108 and #1107 I want to reduce allocations in selectServerFromDescription func which flat allocations in my case are > 100GB and which possess the very first place by flat allocations in my long running job.

Benchmark attached without changes to the code:
```
BenchmarkSelectServerFromDescription
BenchmarkSelectServerFromDescription/AllFit
BenchmarkSelectServerFromDescription/AllFit-10         	   63248	     17195 ns/op	  111265 B/op	       8 allocs/op
BenchmarkSelectServerFromDescription/AllButOneFit
BenchmarkSelectServerFromDescription/AllButOneFit-10   	   65990	     18173 ns/op	  111265 B/op	       8 allocs/op
BenchmarkSelectServerFromDescription/HalfFit
BenchmarkSelectServerFromDescription/HalfFit-10        	  110966	      9053 ns/op	   53920 B/op	       7 allocs/op
BenchmarkSelectServerFromDescription/OneFit
BenchmarkSelectServerFromDescription/OneFit-10         	 3362158	       351.3 ns/op	     416 B/op	       1 allocs/op
PASS
```

And with changes:
```
BenchmarkSelectServerFromDescription
BenchmarkSelectServerFromDescription/AllFit
BenchmarkSelectServerFromDescription/AllFit-10         	 2548441	       480.4 ns/op	     896 B/op	       1 allocs/op
BenchmarkSelectServerFromDescription/AllButOneFit
BenchmarkSelectServerFromDescription/AllButOneFit-10   	  159066	      6740 ns/op	   41856 B/op	       2 allocs/op
BenchmarkSelectServerFromDescription/HalfFit
BenchmarkSelectServerFromDescription/HalfFit-10        	  295821	      3710 ns/op	   21376 B/op	       2 allocs/op
BenchmarkSelectServerFromDescription/OneFit
BenchmarkSelectServerFromDescription/OneFit-10         	 1814713	       639.6 ns/op	    1312 B/op	       2 allocs/op
PASS
```