@isopov 2022-10-27
BTW in all the cases it seems beneficial to use bool slice for filtering like
```
	allowedIndexes := make([]bool, len(desc.Servers))
	allowedIndexesCount := 0
	for i, s := range desc.Servers {
		if s.Kind != description.Unknown {
			allowedIndexes[i] = true
			allowedIndexesCount++
		}
	}
	var allowed []description.Server
	if allowedIndexesCount == len(desc.Servers) {
		allowed = desc.Servers
	} else {
		allowed = make([]description.Server, 0, allowedIndexesCount)
		for i, allowedIndex := range allowedIndexes {
			if allowedIndex {
				allowed = append(allowed, desc.Servers[i])
			}
		}
	}
```
In this case benchmark results are:
```
BenchmarkSelectServerFromDescription
BenchmarkSelectServerFromDescription/AllFit
BenchmarkSelectServerFromDescription/AllFit-10         	 4538070	       276.1 ns/op	     112 B/op	       1 allocs/op
BenchmarkSelectServerFromDescription/AllButOneFit
BenchmarkSelectServerFromDescription/AllButOneFit-10   	  146278	      7216 ns/op	   41072 B/op	       2 allocs/op
BenchmarkSelectServerFromDescription/HalfFit
BenchmarkSelectServerFromDescription/HalfFit-10        	  297627	      3904 ns/op	   20592 B/op	       2 allocs/op
BenchmarkSelectServerFromDescription/OneFit
BenchmarkSelectServerFromDescription/OneFit-10         	 2960940	       393.9 ns/op	     528 B/op	       2 allocs/op
PASS
```

Should I use this approach in all three places?